requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
kombu==5.3.5
prompt-toolkit==3.0.43
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg-pool==3.2.1
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2024.1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
websockets==12.0
